requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
streamlit>=1.28.0
requests>=2.31.0
groq>=0.4.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
